Updated to Lesson 8, Client Side Routing
